requirements.txt
Django>=3.2.4,<=3.3
djangorestframework>=3.12.4,<3.13
psycopg2-binary==2.9.9
drf-spectacular>=0.15.1,<0.16
django-auditlog==3.0.0
reportlab==4.2.0
django-admin-rangefilter==0.13.0
python-dotenv == 1.0.1
django-cors-headers==4.4.0
djangorestframework_simplejwt==5.3.1
django-storages[azure]==1.14.4
channels==3.0.4